pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "fastapi-app"
version = "0.1.0"
description = "Production FastAPI Application"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "sqlmodel>=0.0.24",
    "asyncpg>=0.30.0",
    "redis>=5.2.0",
    "python-multipart>=0.0.12",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "alembic>=1.16.5",
    "pydantic-settings>=2.10.1",
    "uvicorn[standard]>=0.35.0",
    "openai>=1.51.0",
    "tiktoken>=0.8.0",
    "stripe>=11.1.0",
    "cryptography>=42.0.0",
    "celery[redis]>=5.5.3",
    "prometheus-client>=0.20.0",
    "slowapi>=0.1.9",
    "tenacity>=8.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-alembic>=0.4.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "pre-commit>=3.8.0",
    "pytest-mock>=3.14.0",
    "respx>=0.21.0",
    "factory-boy>=3.3.1",
]

prod = [
    "gunicorn>=23.0.0",
    "uvicorn-worker>=0.3.0",
    "aiosqlite>=0.21.0",
]

finance = [
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "scipy>=1.14.0",
]

privacy = [
    "data-anonymization>=2.1.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --asyncio-mode=auto"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "alembic: marks tests as alembic migration tests",
]

[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
]
ignore = ["E501", "B008", "UP007"]  # Line too long, Depends default, Union syntax

[tool.ruff.isort]
known-first-party = ["src"]
force-sort-within-sections = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports in __init__.py
"tests/**/*" = ["S101"]   # Use of assert detected